feat: Add inputmodal
