+++ src/components/DefaultButton/index.js
@@ -1,4 +1,5 @@
 import React from "react";
+
 import { TouchableOpacity, Text, StyleSheet } from "react-native";
 
 /*
