+++ src/components/index.js
@@ -1,6 +1,15 @@
 import DefaultButton from "./DefaultButton";
-import SquareButton from "./SquareButton";
-import SmallButton from "./SamllButton";
+import InputModal from "./InputModal/InputModal";
+import InputModalContainer from "./InputModal/InputModalContainer";
 import Logo from "./Logo";
+import SmallButton from "./SamllButton";
+import SquareButton from "./SquareButton";
 
-export { DefaultButton, SquareButton, SmallButton, Logo };
+export {
+  DefaultButton,
+  SquareButton,
+  SmallButton,
+  Logo,
+  InputModalContainer,
+  InputModal,
+};
